# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "QSOFT test API",
        "description": "Документация API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get user list",
                "operationId": "325fd3d55bbe9bb39645269928fa4983",
                "responses": {
                    "200": {
                        "description": "Success, redirect to order create"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create new order",
                "operationId": "445bab4f7dc9ea6fcd54f286c0209478",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "message"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "user"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "some text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success, redirect to order create"
                    }
                }
            }
        },
        "/orders/{id}": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Set comment",
                "description": "Set status Resolved, and add to order comment",
                "operationId": "05db5a94387077dbc228ff3dc046abe0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "comment"
                                ],
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "order comment text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Sussecc operation"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Orders"
        }
    ]
}